improve separation of concern
