+++ UwpAspNetCore/Startup.cs
@@ -6,20 +6,19 @@ using Microsoft.AspNetCore.Mvc.ApplicationParts;
 using Microsoft.Extensions.Configuration;
 using Microsoft.Extensions.DependencyInjection;
 using System.Reflection;
+using UwpAspNetLib;
 using UwpAspNetLib.Hubs;
 
 namespace UwpAspNetCore
 {
-    class Startup
+    class UwpStartup : Startup
     {
-        public Startup(IConfiguration configuration)
+        public UwpStartup(IConfiguration configuration)
+            :base(configuration)
         {
-            Configuration = configuration;
         }
-
-        public IConfiguration Configuration { get; }
-
-        public void ConfigureServices(IServiceCollection services)
+        
+        public override void ConfigureServices(IServiceCollection services)
         {
             var manager = new ApplicationPartManager();
             var assemblies = new Assembly[] { Assembly.GetExecutingAssembly(), Assembly.Load("UwpAspNetLib"), Assembly.Load("UwpAspNetLib.Views") };
@@ -36,45 +35,7 @@ namespace UwpAspNetCore
 
             services.AddSingleton(manager);
 
-            //services.Configure<CookiePolicyOptions>(options =>
-            //{
-            //    // This lambda determines whether user consent for non-essential cookies is needed for a given request.
-            //    options.CheckConsentNeeded = context => true;
-            //    options.MinimumSameSitePolicy = SameSiteMode.None;
-            //});
-
-            services.AddSignalR();
-
-            services.AddMvc().SetCompatibilityVersion(CompatibilityVersion.Version_2_1);
-        }
-
-        public void Configure(IApplicationBuilder app, IHostingEnvironment env)
-        {
-            if (env.IsDevelopment())
-            {
-                app.UseDeveloperExceptionPage();
-            }
-            else
-            {
-                app.UseExceptionHandler("/Home/Error");
-                //app.UseHsts();
-            }
-
-            //app.UseHttpsRedirection();
-            app.UseStaticFiles();
-            //app.UseCookiePolicy();
-
-            app.UseSignalR(routes =>
-            {
-                routes.MapHub<Chat>("/chat");
-            });
-
-            app.UseMvc(routes =>
-            {
-                routes.MapRoute(
-                    name: "default",
-                    template: "{controller=Home}/{action=Index}/{id?}");
-            });
+            base.ConfigureServices(services);
         }
     }
 }
